Tabular Listings of GeoRegions

A list of GeoRegions and their basic properties can be called using the function tableGeoRegions(), which lists all available GeoRegions, both predefined and custom/user-defined.

julia> using GeoRegions
julia> resetGeoRegions(allfiles=true)[ Info: 2022-07-30T01:58:33.356 - GeoRegions.jl - Resetting both the master and custom lists of GeoRegions back to the default
┌ Warning: 2022-07-30T01:58:33.356 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/rectlist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/rectlist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:33.356 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/polylist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/polylist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:33.356 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/giorgi.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/giorgi.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:33.356 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/srex.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/srex.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:33.356 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/ar6.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/ar6.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
julia> tableGeoRegions(onlycustom=true) ----- ------------ ------- -------- --------------
  ID       Type      Name    Parent       File
 ----- ------------ ------- -------- --------------
  GLB   RectRegion   Globe    GLB     rectlist.txt
 ----- ------------ ------- -------- --------------

GeoRegions.tableGeoRegions — Method

tableGeoRegions(; onlycustom::Bool) -> nothing

Display all available GeoRegions in tabular format.

Keyword Arguments

onlycustom : If true, only custom, user-defined GeoRegions will be displayed. To show all available GeoRegions (including predefined GeoRegions), set onlycustom to false

More specific lists of GeoRegions

Besides listing all available GeoRegions, it is also possible to list a variety of different specifications of GeoRegions with the tableGeoRegions() function, as well as tableRectRegions() and tablePolyRegions().

GeoRegions.tableRectRegions — Function

tableRectRegions(;
    custom :: Bool = true,
    giorgi :: Bool = false
) -> nothing

Display all available RectRegions in tabular format.

Keyword Arguments

custom : If true, display custom user-defined GeoRegions. Default is true

giorgi : If true, display Giorgi predefined GeoRegions. Default is false

GeoRegions.tablePolyRegions — Function

tablePolyRegions(;
    custom :: Bool = true,
    srex :: Bool = false,
    ar6  :: Bool = false
) -> nothing

Display all available PolyRegions in tabular format.

Keyword Arguments

custom : If true, display custom user-defined GeoRegions. Default is true

srex : If true, display SREX predefined GeoRegions. Default is false

ar6 : If true, display IPCC AR6 predefined GeoRegions. Default is false

Listing GeoRegions inside a Custom File

Similarly, it is also possible to list the GeoRegions in custom files. The template for these files can be retrieved using the function templateGeoRegions()

GeoRegions.tableGeoRegions — Method

tableGeoRegions(fname::AbstractString) -> nothing

Display all available GeoRegions in tabular format.

Keyword Arguments

onlycustom : If true, only custom, user-defined GeoRegions will be displayed. To show all available GeoRegions (including predefined GeoRegions), set onlycustom to false

Master List of All Predefined GeoRegions

julia> using GeoRegions
julia> resetGeoRegions(allfiles=true)[ Info: 2022-07-30T01:58:34.891 - GeoRegions.jl - Resetting both the master and custom lists of GeoRegions back to the default
┌ Warning: 2022-07-30T01:58:34.891 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/rectlist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/rectlist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:34.892 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/polylist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/polylist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:34.892 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/giorgi.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/giorgi.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:34.892 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/srex.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/srex.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2022-07-30T01:58:34.892 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/ar6.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/ar6.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
julia> tableGeoRegions() ---------- ------------ ---------------------------- -------- --------------
     ID         Type      Name                         Parent       File
 ---------- ------------ ---------------------------- -------- --------------
    GLB      RectRegion   Globe                         GLB     rectlist.txt
   GF_AUS    RectRegion   Australia                     GLB      giorgi.txt
   GF_AMZ    RectRegion   Amazon Basin                  GLB      giorgi.txt
   GF_SSA    RectRegion   Southern South America        GLB      giorgi.txt
   GF_CAM    RectRegion   Central America               GLB      giorgi.txt
   GF_WNA    RectRegion   Western North America         GLB      giorgi.txt
   GF_CNA    RectRegion   Central North America         GLB      giorgi.txt
   GF_ENA    RectRegion   Eastern North America         GLB      giorgi.txt
   GF_ALA    RectRegion   Alaska                        GLB      giorgi.txt
   GF_GRL    RectRegion   Greenland                     GLB      giorgi.txt
   GF_MED    RectRegion   Mediterranean Basin           GLB      giorgi.txt
   GF_NEU    RectRegion   Northern Europe               GLB      giorgi.txt
   GF_WAF    RectRegion   Western Africa                GLB      giorgi.txt
   GF_EAF    RectRegion   Eastern Africa                GLB      giorgi.txt
   GF_SAF    RectRegion   Southern Africa               GLB      giorgi.txt
   GF_SAH    RectRegion   Sahara                        GLB      giorgi.txt
   GF_SEA    RectRegion   Southeast Asia                GLB      giorgi.txt
   GF_EAS    RectRegion   East Asia                     GLB      giorgi.txt
   GF_SAS    RectRegion   South Asia                    GLB      giorgi.txt
   GF_CAS    RectRegion   Central Asia                  GLB      giorgi.txt
   GF_TIB    RectRegion   Tibet                         GLB      giorgi.txt
   GF_NAS    RectRegion   North Asia                    GLB      giorgi.txt
  SRX_ALA    PolyRegion   Alaska/N.W. Canada            GLB       srex.txt
  SRX_AMZ    PolyRegion   Amazon                        GLB       srex.txt
  SRX_CAM    PolyRegion   Central America/Mexico        GLB       srex.txt
  SRX_CAR    PolyRegion   Caribbean                     GLB       srex.txt
  SRX_CAS    PolyRegion   Central Asia                  GLB       srex.txt
  SRX_CEU    PolyRegion   Central Europe                GLB       srex.txt
  SRX_CGI    PolyRegion   Canada/Greenland/Iceland      GLB       srex.txt
  SRX_CNA    PolyRegion   Central North America         GLB       srex.txt
  SRX_EAF    PolyRegion   East Africa                   GLB       srex.txt
  SRX_EAS    PolyRegion   East Asia                     GLB       srex.txt
  SRX_ENA    PolyRegion   East North America            GLB       srex.txt
  SRX_MED    PolyRegion   South Europe/Mediterranean    GLB       srex.txt
  SRX_NAS    PolyRegion   North Asia                    GLB       srex.txt
  SRX_NAU    PolyRegion   North Australia               GLB       srex.txt
  SRX_NEB    PolyRegion   Northeast Brazil              GLB       srex.txt
  SRX_NEU    PolyRegion   North Europe                  GLB       srex.txt
  SRX_SAF    PolyRegion   Southern Africa               GLB       srex.txt
  SRX_SAH    PolyRegion   Sahara                        GLB       srex.txt
  SRX_SAS    PolyRegion   South Asia                    GLB       srex.txt
  SRX_SAU    PolyRegion   South Australia/NewZealand    GLB       srex.txt
  SRX_SEA    PolyRegion   Southeast Asia                GLB       srex.txt
  SRX_SSA    PolyRegion   Southeastern South America    GLB       srex.txt
  SRX_TIB    PolyRegion   Tibetan Plateau               GLB       srex.txt
  SRX_WAF    PolyRegion   West Africa                   GLB       srex.txt
  SRX_WAS    PolyRegion   West Asia                     GLB       srex.txt
  SRX_WNA    PolyRegion   Northwest America             GLB       srex.txt
  SRX_WSA    PolyRegion   West Coast South America      GLB       srex.txt
  SRX_ANT    PolyRegion   Antarctica                    GLB       srex.txt
  SRX_ARC    PolyRegion   Arctic                        GLB       srex.txt
  SRX_NTP    PolyRegion   Pacific Islands               GLB       srex.txt
  SRX_STP    PolyRegion   Southern Tropical Pacific     GLB       srex.txt
  SRX_ETP    PolyRegion   Eastern Tropical Pacific      GLB       srex.txt
  SRX_WIO    PolyRegion   West Indian Ocean             GLB       srex.txt
  AR6_GIC    PolyRegion   Greenland/Iceland             GLB       ar6.txt
  AR6_NWN    PolyRegion   Northwest North America       GLB       ar6.txt
  AR6_NEN    PolyRegion   Northeast North America       GLB       ar6.txt
  AR6_WNA    PolyRegion   Western North America         GLB       ar6.txt
  AR6_CNA    PolyRegion   Central North America         GLB       ar6.txt
  AR6_ENA    PolyRegion   Eastern North America         GLB       ar6.txt
  AR6_NCA    PolyRegion   Northern Central America      GLB       ar6.txt
  AR6_SCA    PolyRegion   Southern Central America      GLB       ar6.txt
  AR6_CAR    PolyRegion   Caribbean                     GLB       ar6.txt
  AR6_NWS    PolyRegion   Northwest South America       GLB       ar6.txt
  AR6_NSA    PolyRegion   Northern South America        GLB       ar6.txt
  AR6_NES    PolyRegion   Northeast South America       GLB       ar6.txt
  AR6_SAM    PolyRegion   South American Monsoon        GLB       ar6.txt
  AR6_SWS    PolyRegion   Southwest South America       GLB       ar6.txt
  AR6_SES    PolyRegion   Southeast South America       GLB       ar6.txt
  AR6_SSA    PolyRegion   Southern South America        GLB       ar6.txt
  AR6_NEU    PolyRegion   Northern Europe               GLB       ar6.txt
  AR6_WCE    PolyRegion   West and Central Europe       GLB       ar6.txt
  AR6_EEU    PolyRegion   Eastern Europe                GLB       ar6.txt
  AR6_MED    PolyRegion   Mediterranean                 GLB       ar6.txt
  AR6_SAH    PolyRegion   Sahara                        GLB       ar6.txt
  AR6_WAF    PolyRegion   Western Africa                GLB       ar6.txt
  AR6_CAF    PolyRegion   Central Africa                GLB       ar6.txt
  AR6_NEAF   PolyRegion   Northern East Africa          GLB       ar6.txt
  AR6_SEAF   PolyRegion   Southern East Africa          GLB       ar6.txt
  AR6_WSAF   PolyRegion   Western South Africa          GLB       ar6.txt
  AR6_ESAF   PolyRegion   Eastern South Africa          GLB       ar6.txt
  AR6_MDG    PolyRegion   Madagascar                    GLB       ar6.txt
  AR6_RAR    PolyRegion   Arctic Russia                 GLB       ar6.txt
  AR6_WSB    PolyRegion   Western Siberia               GLB       ar6.txt
  AR6_ESB    PolyRegion   Eastern Siberia               GLB       ar6.txt
  AR6_RFE    PolyRegion   Far East Russia               GLB       ar6.txt
  AR6_WCA    PolyRegion   Western Central Africa        GLB       ar6.txt
  AR6_ECA    PolyRegion   Eastern Central Africa        GLB       ar6.txt
  AR6_TIB    PolyRegion   Tibetian Plateau              GLB       ar6.txt
  AR6_EAS    PolyRegion   East Asia                     GLB       ar6.txt
  AR6_ARP    PolyRegion   Arabian Peninsular            GLB       ar6.txt
  AR6_SAS    PolyRegion   South Asia                    GLB       ar6.txt
  AR6_SEA    PolyRegion   Southeast Asia                GLB       ar6.txt
  AR6_NAU    PolyRegion   Northern Australia            GLB       ar6.txt
  AR6_CAU    PolyRegion   Central Australia             GLB       ar6.txt
  AR6_EAU    PolyRegion   Eastern Australia             GLB       ar6.txt
  AR6_SAU    PolyRegion   Southern Australia            GLB       ar6.txt
   AR6_NZ    PolyRegion   New Zealand                   GLB       ar6.txt
  AR6_EAN    PolyRegion   East Antarctica               GLB       ar6.txt
  AR6_WAN    PolyRegion   West Antarctica               GLB       ar6.txt
  AR6_ARO    PolyRegion   Arctic Ocean                  GLB       ar6.txt
  AR6_NPO    PolyRegion   Northern Pacific Ocean        GLB       ar6.txt
  AR6_EPO    PolyRegion   Equatorial Pacific Ocean      GLB       ar6.txt
  AR6_SPO    PolyRegion   Southern Pacific Ocean        GLB       ar6.txt
  AR6_NAO    PolyRegion   Northern Atlantic Ocean       GLB       ar6.txt
  AR6_EAO    PolyRegion   Equatorial Atlantic Ocean     GLB       ar6.txt
  AR6_SAO    PolyRegion   Southern Atlantic Ocean       GLB       ar6.txt
  AR6_ARS    PolyRegion   Arabian Sea                   GLB       ar6.txt
  AR6_BOB    PolyRegion   Bay of Bengal                 GLB       ar6.txt
  AR6_EIO    PolyRegion   Equatorial Indian Ocean       GLB       ar6.txt
  AR6_SIO    PolyRegion   Southern Indian Ocean         GLB       ar6.txt
  AR6_SOO    PolyRegion   Southern Ocean                GLB       ar6.txt
 ---------- ------------ ---------------------------- -------- --------------