Land-Sea Mask Datasets

Starting from v5, GeoRegions.jl now fully supports the retrieval of 30- and 60-arcsecond ETOPO orographic data, which includes the following options:

Topography that follows the ice-surface (surface)

Topography that follows the bedrock beneath the ice (bedrock)

Geoid data that converts the Topography into WTGS84 ellipsoid elevation heights

The AbstractLandSea SuperType

The data is loaded into a LandSea dataset struct-type, which is a subset of the supertype AbstractLandSea. We export this supertype, and its abstract subtypes LandSeaTopo and LandSeaFlat, so that other packages can extend the use of these abstract types with their own LandSea Types.

For examples of how to extend the functionality of the LandSea datasets, you may refer to the following examples:

NASAPrecipitation.jl

ERA5Reanalysis.jl

GeoRegions.AbstractLandSea — Type

AbstractLandSea

Abstract supertype for LandSea Datasets, with the following subtypes:

LandSeaTopo <: AbstractLandSea
LandSeaFlat <: AbstractLandSea

Both LandSeaTopo and LandSeaFlat types contain the following fields:

lon - Vector containing the longitude points for the Land-Sea Dataset

lat - Vector containing the latitude points for the Land-Sea Dataset

lsm - Array containing data regarding the Land-Sea Mask. 1 is Land, 0 is Ocean, NaN is outside the bounds of the GeoRegion

mask - Mask determining if point is within the GeoRegion or not. 1 is true, 0 is false.

A LandSeaTopo type will also contain the following field:

z - Array containing data regarding the Orographic Height in meters. NaN is outside the bounds of the GeoRegion

The LandSea Type for GeoRegions.jl

GeoRegions.LandSea — Type

GeoRegion.LandSea <: LandSeaTopo

Object containing information on the ETOPO 2022 Land Sea mask for a GeoRegion.