Tabular Listings of GeoRegions

A list of GeoRegions and their basic properties can be called using the function tableGeoRegions(), which lists all available GeoRegions, both predefined and custom/user-defined.

julia> using GeoRegions
julia> resetGeoRegions(allfiles=true)[ Info: 2023-05-28T04:01:43.106 - GeoRegions.jl - Resetting both the master and custom lists of GeoRegions back to the default
┌ Warning: 2023-05-28T04:01:43.106 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/rectlist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/rectlist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2023-05-28T04:01:43.113 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/polylist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/polylist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2023-05-28T04:01:43.114 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/giorgi.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/giorgi.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2023-05-28T04:01:43.114 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/srex.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/srex.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
┌ Warning: 2023-05-28T04:01:43.114 - GeoRegions.jl - Overwriting /home/runner/.julia/files/GeoRegions/ar6.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/ar6.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:417
julia> tableGeoRegions(onlycustom=true) ----- ------------ ------- -------- --------------
  ID       Type      Name    Parent       File
 ----- ------------ ------- -------- --------------
  GLB   RectRegion   Globe    GLB     rectlist.txt
 ----- ------------ ------- -------- --------------

GeoRegions.tableGeoRegions — Method

tableGeoRegions(; onlycustom::Bool) -> nothing

Display all available GeoRegions in tabular format.

Keyword Arguments

onlycustom : If true, only custom, user-defined GeoRegions will be displayed. To show all available GeoRegions (including predefined GeoRegions), set onlycustom to false

More specific lists of GeoRegions

Besides listing all available GeoRegions, it is also possible to list a variety of different specifications of GeoRegions with the tableGeoRegions() function, as well as tableRectRegions() and tablePolyRegions().

GeoRegions.tableRectRegions — Function

tableRectRegions(;
    custom :: Bool = true,
    giorgi :: Bool = false
) -> nothing

Display all available RectRegions in tabular format.

Keyword Arguments

custom : If true, display custom user-defined GeoRegions. Default is true

giorgi : If true, display Giorgi predefined GeoRegions. Default is false

GeoRegions.tablePolyRegions — Function

tablePolyRegions(;
    custom :: Bool = true,
    srex :: Bool = false,
    ar6  :: Bool = false
) -> nothing

Display all available PolyRegions in tabular format.

Keyword Arguments

custom : If true, display custom user-defined GeoRegions. Default is true

srex : If true, display SREX predefined GeoRegions. Default is false

ar6 : If true, display IPCC AR6 predefined GeoRegions. Default is false

Listing GeoRegions inside a Custom File

Similarly, it is also possible to list the GeoRegions in custom files. The template for these files can be retrieved using the function templateGeoRegions()

GeoRegions.tableGeoRegions — Method

tableGeoRegions(fname::AbstractString) -> nothing

Display all available GeoRegions in tabular format.

Keyword Arguments

onlycustom : If true, only custom, user-defined GeoRegions will be displayed. To show all available GeoRegions (including predefined GeoRegions), set onlycustom to false