Is Point/Grid in Region?

When dealing with geographic data, we often wish to check if a point or the bounds of a grid is within a given region. GeoRegions.jl allows you to perform this check easily with ispointinregion and isgridinregion.

This page goes through the following:

Ways to express a Region

How to use ispointinregion()

How to use ispointinregion()

Expressing a Region

In both ispointinregion() and ispointinregion(), the region can be expressed in the form of the following:

Grid Boundaries regionbounds which is a vector taking the form of [N,S,E,W] coordinates

A pair of longitude and latitude vectors, rlon and rlat

A GeoRegion ID

Thresholds

You might have noticed that for methods (1) and (3), there are variables defining the threshold of error tlon and tlat. For method (2), these variables are defined to be half the steps in rlon and rlat respectively. For example, if rlon = collect(18:25) and rlat = collect(0.1:0.1:10), then tlon and tlat are given by:

tlon = (rlon[2]-rlon[1])/2 = 0.5;
tlat = (rlat[2]-rlat[1])/2 = 0.05;

Note that for method 2, it is assumed that the points in rlon and rlat are evenly spaced.

The first two methods are provided in order for the functions to be generally applicable when a GeoRegion has not been explicitly created for this particular region.

An example of each of the three different methods is as follows:

Using the regionbounds method:

julia> plon = 79.5; plat = 1.76; regionbounds = [3,0,80,79];
julia> ispointinregion(plon,plat,regionbounds)
true

Using rlon and rlat:

julia> plon = 79.5; rlon = collect(70:95);
julia> plat = 1.76; rlat = collect(0:5);
julia> ispointinregion(plon,plat,rlon,rlat)
true

Using GeoRegion ID

julia> ID = "GF_WAF"; #regionbounds =  [20,-15,25,340]
julia> plon = 100; plat = 1.76; ispointinregion(plon,plat,ID)
ERROR: Requested coordinates [100,1.76] are not within the specified region boundaries.

julia> plon = 0;   plat = 1.76; ispointinregion(plon,plat,ID)
true

Is Point in Region?

There are six different methods of querying if a point is in a given region using ispointinregion. This is a combination of having three different methods to express the region (as mentioned above), and two different methods of expressing a point.

The two different methods of expression a point are:

a pair of longitude and latitude points plon and plat

a coordinate vector pcoord = (plon,plat)

Examples

Text

Methods

GeoRegions.ispointinregion — Function

ispointinregion(plon, plat, rlon, rlat; throw=true) -> Bool

Check if a point with coordinates (plon,plat) is found within a region defined by the longitude and latitude vectors rlon and rlat

Arguments:

plon::Real : Longitude of the point in question.

plat::Real : Latitude of the point in question.

rlon::Vector{<:Real} : Longitude vector spanning the region. Points should be evenly spaced.

rlat::Vector{<:Real} : Latitude vector spanning the region. Points should be evenly spaced.

Keyword Arguments:

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

ispointinregion(plon, plat, regionbounds; tlon=0, tlat=0, throw=true) -> Bool

Check if a point with coordinates (plon,plat) is found within a region defined by the bounds regionbounds in [N,S,E,W] format.

Arguments:

plon::Real : Longitude of the point in question.

plat::Real : Latitude of the point in question.

regionbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the region.

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

ispointinregion(plon, plat, gregID; tlon=0, tlat=0, throw=true) -> Bool

Check if a point with coordinates pcoord = [plon,plat] is found within a predefined GeoRegion.

Arguments:

plon::Real : Longitude of the point in question.

plat::Real : Latitude of the point in question.

gregID::AbstractString : GeoRegion ID

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

ispointinregion(pcoord, rlon, rlat; throw=true) -> Bool

Check if a point with coordinates pcoord = [plon,plat] is found within a region defined by the longitude and latitude vectors rlon and rlat.

Arguments:

pcoord::Vector{<:Real} : Vector defining coordinates of point in [plon,plat]

rlon::Vector{<:Real} : Longitude vector spanning the region. Points should be evenly spaced.

rlat::Vector{<:Real} : Latitude vector spanning the region. Points should be evenly spaced.

Keyword Arguments:

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

ispointinregion(pcoord, regionbounds; tlon=0, tlat=0, throw=true) -> Bool

Check if a point with coordinates pcoord = [plon,plat] is found within a region defined by the bounds regionbounds in [N,S,E,W] format.

Arguments:

pcoord::Vector{<:Real} : Vector defining coordinates of point in [plon,plat]

regionbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the region.

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

ispointinregion(pcoord, gregID; tlon=0, tlat=0, throw=true) -> Bool

Check if a point with coordinates pcoord = [plon,plat] is found within a predefined GeoRegion.

Arguments:

pcoord::Vector{<:Real} : Vector defining coordinates of point in [plon,plat]

gregID::AbstractString : GeoRegion ID

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

Is Grid in Regions?

There are three different methods of querying if a point is in a given region using isgridinregion, because of the three different methods to express the region (as mentioned above).

Examples

Text

Methods

GeoRegions.isgridinregion — Function

isgridinregion(gridbounds, regionbounds; tlon=0, tlat=0, throw=true) -> Bool

Check if a grid defined by gridbounds = [gN,gS,gE,gW] is found within a predefined GeoRegion.

Arguments:

gridbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the grid in question.

regionbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the region.

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

isgridinregion(gridbounds, rlon, rlat; throw=true) -> Bool

Check if a grid defined by gridbounds = [gN,gS,gE,gW] is found within a predefined GeoRegion.

Arguments:

gridbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the grid in question.

rlon::Vector{<:Real} : Longitude vector spanning the region. Points should be evenly spaced.

rlat::Vector{<:Real} : Latitude vector spanning the region. Points should be evenly spaced.

Keyword Arguments:

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

isgridinregion(gridbounds, gregID; tlon=0, tlat=0, throw=true) -> Bool

Check if a grid defined by gridbounds = [gN,gS,gE,gW] is found within a predefined GeoRegion.

Arguments:

gridbounds::Vector{<:Real} : Vector defining the [North,South,East,West] bounds of the grid in question.

gregID::AbstractString : GeoRegion ID

Keyword Arguments:

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.