Application Programming Interface (APIs)

This page contains a brief description of the functions not covered in the tutorials. I aim to describe what the functions do, their arguments, and what is returned.

GeoRegions.checkpoint — Function

checkpoint(plon, plat, rN, rS, rE, rW, tlon, tlat, throw) -> Bool

Check if a point with longitude and latitude coordinates plon and plat is found within a region defined by the bounds [rN,rS,rE,rW] format, with tlon and tlat being the threshold/margin of error for the longitude and latitude respectively.

Arguments:

plon::Real : Longitude of the point in question.

plat::Real : Latitude of the point in question.

rN::Real : North bound of region

rS::Real : South bound of region

rE::Real : East bound of region

rW::Real : West bound of region

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.

GeoRegions.checkgrid — Function

checkgrid(gN, gS, gE, gW, rN, rS, rE, rW, tlon, tlat, throw) -> Bool

Check if a grid with bounds [gN,gS,gE,gW] is found within a region defined by the bounds [rN,rS,rE,rW] format, with tlon and tlat being the threshold/margin of error for the longitude and latitude respectively.

Arguments:

gN::Real : North bound of grid

gS::Real : South bound of grid

gE::Real : East bound of grid

gW::Real : West bound of grid

rN::Real : North bound of region

rS::Real : South bound of region

rE::Real : East bound of region

rW::Real : West bound of region

tlon::Real : Threshold for longitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the longitude bounds

tlat::Real : Threshold for latitude bounds in °. If set to 0, means that there is no leniency for the point to fall outside the latitude bounds

throw::Bool : If throw=true, then if (plon,plat) is not within the region, an error is thrown and the program stops running.