Creating GeoRegions

In this section, we go through the basic steps of creating a GeoRegion, and removing it once defined so it can be redefined.

Defining a new GeoRegion

We use the functions RectRegion and PolyRegion to define rectilinear and polygonal regions respectively. For example, we construct the sample RectRegion TRC

julia> using GeoRegions

julia> RectRegion("TRC","GLB","Test Rectangle Region",[30,20,50,10])
┌ Warning: 2021-06-25T02:55:54.262 - TRC is not a valid GeoRegion.  Use either RectRegion() or PolyRegion() to add this GeoRegion to the list.
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:405
[ Info: 2021-06-25T02:55:54.355 - Adding the GeoRegion TRC to the list.
The Rectilinear Region TRC has the following properties:
    Region ID (regID) : TRC
    Parent ID (parID) : GLB
    Name      (name)  : Test Rectangle Region
    Bounds  (N,S,E,W) : [30.0, 20.0, 50.0, 10.0]
        (is180,is360) : (false, true)

julia> PolyRegion("TPL","GLB","Test Polygonal Region",[30,40,50,40,30],[20,30,20,10,20])
┌ Warning: 2021-06-25T02:55:55.035 - TPL is not a valid GeoRegion.  Use either RectRegion() or PolyRegion() to add this GeoRegion to the list.
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:405
[ Info: 2021-06-25T02:55:55.035 - Adding the GeoRegion TPL to the list.
The Polygonal Region TPL has the following properties:
    Region ID (regID) : TPL
    Parent ID (parID) : GLB
    Name       (name) : Test Polygonal Region
    Bounds  (N,S,E,W) : [30.0, 10.0, 50.0, 30.0]
    Shape     (shape) : Point2{Float64}[[30.0, 20.0], [40.0, 30.0], [50.0, 20.0], [40.0, 10.0], [30.0, 20.0]]
        (is180,is360) : (false, true)

Defining PolyRegions

When defining PolyRegions, the first and last set of (lon,lat) coordinates must be the same (i.e. a closed polygon must be defined)

The GeoRegions will be automatically added to the following files in joinpath(DEPOT_PATH[1],"files","GeoRegions"):

RectRegions will be added to rectlist.txt

PolyRegions will be added to polylist.txt

If the GeoRegion ID TRC already exists, however, this will throw an error

julia> using GeoRegions

julia> RectRegion("TRC","GLB","Test Rectangle Region2",[40,20,52,10])
ERROR: 2021-06-25T02:55:55.403 - The GeoRegion TRC has already been defined.  Please use another identifier.

GeoRegions.RectRegion — Method

RectRegion(
    RegID :: AbstractString,
    ParID :: AbstractString,
    name  :: AbstractString,
    bound :: Vector{<:Real},
    ST = String,
    FT = Float64
) -> RectRegion{ST,FT}

Creates a rectilinear GeoRegion RegID.

Arguments

RegID : The keyword ID that will be used to identify the GeoRegion. If the ID is already in use, then an error will be thrown.

ParID : The ID of the parent GeoRegion where information can be extracted from

name : A name for the GeoRegion (meta information, can be used in Logging)

bound : The [N,S,E,W] coordinates defining the region

GeoRegions.PolyRegion — Method

PolyRegion(
    RegID :: AbstractString,
    ParID :: AbstractString,
    name  :: AbstractString,
    lonpt :: Vector{<:Real},
    latpt :: Vector{<:Real},
    ST = String,
    FT = Float64
) -> PolyRegion{ST,FT}

Creates a rectilinear GeoRegion RegID.

Arguments

RegID : The keyword ID that will be used to identify the GeoRegion. If the ID is already in use, then an error will be thrown.

ParID : The ID of the parent GeoRegion where information can be extracted from

name : A name for the GeoRegion (meta information, can be used in Logging)

lonpt : A vector containing the longitude points

latpt : A vector containing the latitude points

Start and End Points

The 1st and last elements of lonpt and latpt must be equal.

Removing an existing GeoRegion

To remove an existing GeoRegion, we can use the function removeGeoRegion

julia> using GeoRegions

julia> removeGeoRegion("TRC")
[ Info: 2021-06-25T02:55:56.290 - Removing the GeoRegion TRC ...

Global GeoRegion

The Global GeoRegion GLB is considered to be an integral part of the GeoRegions.jl package and therefore it cannot be removed.

julia> using GeoRegions

julia> removeGeoRegion("GLB")
ERROR: 2021-06-25T02:55:56.577 - The Global GeoRegion "GLB" is an integral part of the GeoRegions.jl package and cannot be removed.

GeoRegions.removeGeoRegion — Method

removeGeoRegion(RegID::AbstractString)

Creates the GeoRegion associated with the ID RegID.

Arguments

RegID : The keyword ID that will be used to identify the GeoRegion. If the ID is not valid (i.e. not being used), then an error will be thrown.

List of All GeoRegions

If you want to check the list of all GeoRegions that have been defined, use the function listGeoRegions()

julia> using GeoRegions

julia> geovec,filevec,typevec = listGeoRegions();

julia> geovec
114-element Vector{Any}:
 "GLB"
 "TPL"
 "GF_AUS"
 "GF_AMZ"
 "GF_SSA"
 "GF_CAM"
 "GF_WNA"
 "GF_CNA"
 "GF_ENA"
 "GF_ALA"
 ⋮
 "AR6_SPO"
 "AR6_NAO"
 "AR6_EAO"
 "AR6_SAO"
 "AR6_ARS"
 "AR6_BOB"
 "AR6_EIO"
 "AR6_SIO"
 "AR6_SOO"

Reset the list of GeoRegions

Should one wish to entirely reset the list of GeoRegions, one can call resetGeoRegions():

julia> using GeoRegions

julia> resetGeoRegions()
┌ Warning: 2021-06-25T02:55:57.700 - Overwriting /home/runner/.julia/files/GeoRegions/rectlist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/rectlist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341
┌ Warning: 2021-06-25T02:55:57.732 - Overwriting /home/runner/.julia/files/GeoRegions/polylist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/polylist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341

And if one wishes to reset the predefined list of GeoRegions, we can do:

julia> using GeoRegions

julia> resetGeoRegions(allfiles=true)
┌ Warning: 2021-06-25T02:55:57.745 - Overwriting /home/runner/.julia/files/GeoRegions/rectlist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/rectlist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341
┌ Warning: 2021-06-25T02:55:57.745 - Overwriting /home/runner/.julia/files/GeoRegions/polylist.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/polylist.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341
┌ Warning: 2021-06-25T02:55:57.745 - Overwriting /home/runner/.julia/files/GeoRegions/giorgi.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/giorgi.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341
┌ Warning: 2021-06-25T02:55:57.745 - Overwriting /home/runner/.julia/files/GeoRegions/srex.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/srex.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341
┌ Warning: 2021-06-25T02:55:57.745 - Overwriting /home/runner/.julia/files/GeoRegions/ar6.txt with original file in /home/runner/work/GeoRegions.jl/GeoRegions.jl/src/../extra/ar6.txt ...
└ @ GeoRegions ~/work/GeoRegions.jl/GeoRegions.jl/src/Read.jl:341

GeoRegions.listGeoRegions — Function

listGeoRegions()

List all GeoRegions, files the data are stored in, and the Type of GeoRegion they are.

Output

geovec : List of all the GeoRegion IDs

filevec : List of the files that the GeoRegion information is stored in

typevec : List of the Type of the GeoRegion corresponding to the geovec ID list